Initial commit of shortened V3
No pretty cable management for hotend cable connection. Did include cable rings that can screw into the toolhead mount. Includes a blower connector that is printed out of TPU. I'm running this, but its unecessary and needs to be removed. (I'm not sure if the blower fits directly without it, however.)